+++ packages/qb/pqb/src/index.ts
@@ -12,3 +12,4 @@ export * from './errors';
 export * from './relations';
 export * from './columns/operators';
 export { QueryBase } from './queryBase';
+export { testTransaction } from './testTransaction';