+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # test-utils
 
+## 0.0.2
+
+### Patch Changes
+
+- Add testTransaction utility for tests
+
 ## 0.0.1
 
 ### Patch Changes